+++ 1.20/Fabric/src/main/java/fuzs/horseexpert/HorseExpertFabric.java
@@ -1,6 +1,5 @@
 package fuzs.horseexpert;
 
-import fuzs.horseexpert.init.FabricModRegistry;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import net.fabricmc.api.ModInitializer;
 
@@ -8,7 +7,6 @@ public class HorseExpertFabric implements ModInitializer {
 
     @Override
     public void onInitialize() {
-        FabricModRegistry.touch();
         ModConstructor.construct(HorseExpert.MOD_ID, HorseExpert::new);
     }
 }