+++ 1.20/Fabric/src/main/java/fuzs/horseexpert/core/FabricAbstractions.java
@@ -1,6 +1,8 @@
 package fuzs.horseexpert.core;
 
 import dev.emi.trinkets.api.TrinketsApi;
+import fuzs.puzzleslib.api.core.v1.ModLoaderEnvironment;
+import net.minecraft.tags.TagKey;
 import net.minecraft.util.Tuple;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
@@ -11,10 +13,13 @@ import java.util.Optional;
 public class FabricAbstractions implements CommonAbstractions {
 
     @Override
-    public Optional<ItemStack> findEquippedItem(LivingEntity entity, Item item) {
-        return TrinketsApi.getTrinketComponent(entity)
-                .flatMap(component -> component.getEquipped(item).stream()
-                        .findFirst()
-                        .map(Tuple::getB));
+    public Optional<ItemStack> findEquippedItem(LivingEntity entity, TagKey<Item> tagKey) {
+        if (ModLoaderEnvironment.INSTANCE.isModLoaded("trinkets")) {
+            return TrinketsApi.getTrinketComponent(entity).flatMap(component -> component.getEquipped(itemStack -> itemStack.is(tagKey))
+                    .stream().findFirst().map(Tuple::getB))
+                    .or(() -> CommonAbstractions.super.findEquippedItem(entity, tagKey));
+        } else {
+            return CommonAbstractions.super.findEquippedItem(entity, tagKey);
+        }
     }
 }