+++ 1.20/Forge/src/main/java/fuzs/horseexpert/data/ModItemTagsProvider.java
@@ -2,21 +2,24 @@ package fuzs.horseexpert.data;
 
 import fuzs.horseexpert.init.ModRegistry;
 import fuzs.puzzleslib.api.data.v1.AbstractTagProvider;
+import fuzs.puzzleslib.api.init.v3.tags.TypedTagFactory;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 import java.util.concurrent.CompletableFuture;
 
 public class ModItemTagsProvider extends AbstractTagProvider.Items {
 
-    public ModItemTagsProvider(PackOutput packOutput, CompletableFuture<HolderLookup.Provider> lookupProvider, String modId, ExistingFileHelper fileHelper) {
-        super(packOutput, lookupProvider, modId, fileHelper);
+    public ModItemTagsProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override
     protected void addTags(HolderLookup.Provider provider) {
-        this.tag("curios:head").add(ModRegistry.MONOCLE_ITEM.get());
-        this.tag("trinkets:head/face").add(ModRegistry.MONOCLE_ITEM.get());
+        this.tag(ModRegistry.INSPECTION_EQUIPMENT_ITEM_TAG).add(ModRegistry.MONOCLE_ITEM.get());
+        this.tag(TypedTagFactory.ITEM.curios("head")).add(ModRegistry.MONOCLE_ITEM.get());
+        this.tag(TypedTagFactory.ITEM.trinkets("head/face")).add(ModRegistry.MONOCLE_ITEM.get());
     }
 }