+++ 1.20/Forge/src/main/java/fuzs/horseexpert/data/client/ModLanguageProvider.java
@@ -1,15 +1,16 @@
-package fuzs.horseexpert.data;
+package fuzs.horseexpert.data.client;
 
 import fuzs.horseexpert.init.ModRegistry;
 import fuzs.puzzleslib.api.data.v1.AbstractLanguageProvider;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.data.LanguageProvider;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 public class ModLanguageProvider extends AbstractLanguageProvider {
 
-    public ModLanguageProvider(PackOutput packOutput, String modId) {
-        super(packOutput, modId);
+    public ModLanguageProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override