+++ 1.20/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.1.0-1.20.1] - 2023-11-06
+### Changed
+- Both Curios Api and Trinkets are now optional, if none is installed the monocle can be equipped in the helmet armor slot
+- Any equippable item can now provide functionality like the monocle by being added to the `horseexpert:inspection_equipment` item tag
+
 ## [v8.0.1-1.20.1] - 2023-09-06
 ### Fixed
 - Fixed tooltips sometimes showing while riding a mount